pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tesla-mcp-server"
version = "0.1.0"
description = "A Model-Controller-Provider server for Tesla vehicles and solar systems"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "Rob Cerda", email = "me@robcerda.com" }
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.1",
    "pydantic>=2.7.2",
    "python-jose[cryptography]>=3.3.0",
    "mcp[cli]>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "mypy",
]

[project.scripts]
mcp = "tesla_mcp_server.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/tesla_mcp_server"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 